backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pandas>=2.2.0
google-cloud-bigquery==3.13.0
numpy==1.26.2
matplotlib==3.8.2
openai>=1.0.0
python-dotenv>=1.0.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
